requirements.txt
python-telegram-bot==22.1
python-dotenv==1.1.0
colorama==0.4.6
aiohttp==3.12.2